Move CodeEditor to SharedGUI
