+++ DBADashServiceConfig/ServiceConfig.cs
@@ -1,6 +1,5 @@
 ﻿using DBADash;
 using DBADash.Messaging;
-using DBADashGUI.Changes;
 using DBADashGUI.Theme;
 using Humanizer;
 using Microsoft.Data.SqlClient;
