+++ DBADashSharedGUI/CodeEditor/CodeViewer.cs
