+++ DBADashSharedGUI/CodeEditor/CodeViewer.resx
