+++ DBADashSharedGUI/SyntaxHighlighting/PowerShell.xshd
