+++ DBADashSharedGUI/SyntaxHighlighting/SQL-Dark.xshd
