+++ DBADashSharedGUI/SyntaxHighlighting/SQL-Light.xshd
